@@ -4,16 +4,16 @@
   "private": true,
   "dependencies": {
     "@types/jest": "24.0.21",
-    "@types/lodash": "^4.14.144",
+    "@types/lodash": "^4.14.148",
     "@types/node": "12.12.3",
     "@types/react": "16.9.11",
     "@types/react-dom": "16.9.3",
     "@types/styled-components": "^4.4.0",
-    "csv": "^5.1.3",
+    "csv": "^5.3.0",
     "fs": "0.0.1-security",
     "lodash": "^4.17.15",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-scripts": "3.2.0",
     "semantic-ui-react": "^0.88.1",
     "styled-components": "^4.4.1",
