@@ -1,21 +1,19 @@
 import React from "react"
-
-import { BlankSquareContainer, CircleContainer, SquareContainer } from "./style"
+import { BlankSquareContainer, SolvedSquareContainer, UnsolvedSquareContainer } from "./style"
 
 interface SquareProps {
     value: string | null
-    round?: boolean
     onClick?: (e: React.SyntheticEvent) => void
     onMouseOver?: (e: React.SyntheticEvent) => void
-
+    solved: boolean
 }
 
 const Square: React.FunctionComponent<SquareProps> = (props) => {
     const { value, onClick, onMouseOver } = props
-    const SC = value === null ? BlankSquareContainer : props.round ? CircleContainer : SquareContainer
+    const SC = value === null ? BlankSquareContainer : props.solved ? SolvedSquareContainer : UnsolvedSquareContainer
 
     return (
-        <SC onMouseOver={onMouseOver} onClick={onClick}> {value || ""}</SC>
+        <SC onMouseOver={onMouseOver} onClick={onClick}> {value || " "}</SC>
     )
 }
 
